requirements.txt
albumentations
Flask==2.1.0
flask_cors==3.0.10
Flask-Caching
matplotlib
numpy
opencv-python
pandas
pickle5
pillow
scikit-learn
timm
torch==1.11.0 -f https://download.pytorch.org/whl/torch_stable.html
transformers==4.25.1
tqdm